refactor: rename `clear mode` to `clear display`
